pyproject.toml
[tool.poetry]
name = "qpdoc"
version = "0.1.0"
description = "\"Query Processed Documents\""
authors = ["Devel More Group"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.12.1"
pydantic = {version = "^2.0.3", extras = ["email"]}
pydantic-settings = "^2.0.2"
psycopg2-binary = "^2.9.9"
uvicorn = {version = "^0.24.0.post1", extras = ["standard"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
sqlalchemy-utils = "^0.41.1"
python-slugify = "^8.0.1"
fastapi-mail = "^1.4.1"
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"
pypdf2 = "^3.0.1"
langchain = "^0.0.335"
faiss-cpu = "^1.7.4"
openai = "^1.2.4"
pypdf = "^3.17.0"
tiktoken = "^0.5.1"
python-docx = "^1.1.0"
odfpy = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["."]
